src/apps/boards/serializers/comment.py
from rest_framework import serializers

from apps.boards.models import Comment


class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = ('id', 'message',)
