requirements/dev.txt
-r base.txt

pytest-django==4.1.0
mock==4.0.3
factory-boy==3.2.0
django-debug-toolbar

